+++ backend/src/main/java/com/optimize/elykia/core/enumaration/AppNotificationType.java
@@ -3,5 +3,6 @@ package com.optimize.elykia.core.enumaration;
 public enum AppNotificationType {
     PAYMENT_DECLARATION,
     CUSTOMER_ORDER,
-    TONTINE_CATCHUP
+    TONTINE_CATCHUP,
+    TONTINE_PAYMENT_DECLARATION
 }